requirements.txt
Flask
flask-cors
yt-dlp
pytube
requests
python-dotenv
mutagen
redis
celery
gunicorn